// src/themes/themes.d.ts
declare const tailwindThemeClasses: {
    colors: {
        primary: string;
        secondary: string;
        success: string;
        warning: string;
        danger: string;
        info: string;
        dark: string;
        light: string;
        white: string;
    };
    radius: {
        none: string;
        xs: string;
        sm: string;
        md: string;
        lg: string;
        xl: string;
    };
    padding: {
        xl: string;
        lg: string;
        md: string;
        sm: string;
        xs: string;
    };
    sizes: {
        xs: string;
        sm: string;
        md: string;
        lg: string;
        xl: string;
    };
    accentcolor: {
        primary: string;
        secondary: string;
        success: string;
        warning: string;
        danger: string;
        info: string;
        dark: string;
        light: string;
    };
    textcolor: {
        light: string;
        dark: string;
    };
    textsize: {
        xs: string;
        sm: string;
        md: string;
        lg: string;
        xl: string;
    };
};
declare const cssColors: {
    colors: {
        primary: string;
        light: string;
        dark: string;
        info: string;
        danger: string;
        warning: string;
        success: string;
        secondary: string;
    };
};
declare const alignmentClasses: {
    left: string;
    center: string;
    right: string;
};
export { tailwindThemeClasses, alignmentClasses, cssColors };